+++ app/src/main/java/woowacourse/shopping/ui/detail/LastRecentProductUiModel.kt
@@ -1,6 +1,6 @@
 package woowacourse.shopping.ui.detail
 
 class LastRecentProductUiModel(
-    val productId: Long,
+    val productId: Int,
     val title: String,
 )
