+++ app/src/main/java/woowacourse/shopping/ui/detail/OnClickLastRecentProductListener.kt
@@ -1,5 +1,5 @@
 package woowacourse.shopping.ui.detail
 
 fun interface OnClickLastRecentProductListener {
-    fun onClick(productId: Long)
+    fun onClick(productId: Int)
 }
